backend/requirements.txt
chromadb==0.5.3
sentence-transformers==2.7.0
groq==0.9.0
streamlit==1.36.0
pandas==2.2.2
plotly==5.22.0
pyyaml==6.0.1
python-dotenv==1.0.1
posthog==3.7.0
numpy==1.26.4
duckdb==0.10.3
neo4j==5.28.1